Better Ctrl-C handling in the `test --debug` case (#10835)
### Problem
 
Currently when a user hits Ctrl-C while running the `test` goal with the `--debug` flag (which runs test runner process(es) such as `pytest` foregrounded in the terminal using the `InteractiveProcess` mechanism), pants ends up not cleanly exiting the current test runner process, which continues to run in the background. This is due to two slightly-different problems, depending on whether or not pantsd is running. 

If pantsd is not running, then pants successfully ignores the SIGINT signal sent by a terminal Ctrl-C, and forwards that signal to the test process. This is fine if there is exactly one test process being run by the `test` goal; the test runner process will receive the SIGINT, respond to it (in the case of pytest and probably any other reasonable test runner, by exiting), and that is that. However, if there is more than one test source file associated with a test target, or if more than one test target is specified, another test runner subprocess will be started. It is therefore necessary to keep hitting Ctrl-C until all the test runner subprocesses have been manually killed.

If pantsd is running, then pants fails to ignore the SIGINT signal, captures it, and exits, without doing anything to the test runner process, which `InteractiveProcess` has spawned as a separate child process. So, the test runner process continues to run and print data to the terminal until it is finished.

### Solution

In order to fix the no-pantsd case, we add a new flag `forward_signals_to_process` to the `InteractiveProcess` dataclass to make it possible for clients of `InteractiveProcess` to toggle the signal-capturing behavior. We set this flag to False when running debug-mode tests, and leave it as True for other uses of `InteractiveProcess` such as `run` and `repl`, where we do want the subprocess pants is running to handle SIGINT.

In order to fix the pantsd case, we also add logic that sends a SIGINT signal to any subprocesses of pants, when pantsd handles a signal and is shutting down.

### Result

Hitting Ctrl-C when running `test --debug` cleanly exits the running process and gets back to a working terminal.