Backport `--test-extra-env-vars` feature (from #10715) (#10844)
From https://github.com/pantsbuild/pants/pull/10715.

This feature is necessary, understandably, for some users to switch to the v2 test runner. By backporting, we make it a little easier to upgrade from 1.30 to 2.0.

This is not as robust as the 2.0 implementation: we do not have `@_uncachable_rule`, so this is treated like a normal rule. This means that with Pantsd, we won't pick up changes to your environment. You must restart Pantsd (or not use Pantsd).