Fix display of "Snapshotting" in dynamic UI (#10797)
### Problem

A recent change to the string representation of the `Key` type affected the `user_facing_name` of the `Snapshot` variant of `NodeKey`, resulting in the dynamic UI displaying "Snapshotting: PathGlobs" rather than a useful path.

### Solution

Irrespective of the string representation of a, `Key`, the `Snapshot` variant of `NodeKey` doesn't need to contain a `Key` at all; it can contain a `PathGlobs` object. The code that ran a `Snapshot` Node was already parsing the `Key` into a `PathGlobs` (and handling potential errors in that process), so it was easy enough to move that parsing into the location where we create a `Snapshot` from a Python `Value`. Several traits had to be implemented on `PathGlobs` to maintain the trait implementations that `Snapshot` already had.

### Result

Restores the correct behavior for printing the path(s) in a `Snapshot` in the dynamic UI.