Consolidate "runner" PEXes into "tool" PEXes (#10852)
Starting in https://github.com/pantsbuild/pants/pull/9208, we've been using a pattern of separting out a "tool Pex" (the tool's requirements), a "requirements Pex" (user's requirements), and a "runner Pex" (the entry-point). These get combined via `--pex-path`.

There doesn't seem to be a particularly compelling reason to separate the tool Pex from the runner Pex. It's fine to put the entry point on the tool Pex, as that entry point barely ever changes. When tool requirements do change, it's trivial for the new Pex to include the entry point in its setup.

This results in less code and also a less confusing UX, as users no longer have to question why we have both a `mypy.pex` and `mypy_runner.pex`.
 
[ci skip-rust]
[ci skip-build-wheels]