Locate bash before using it to locate BinaryPaths for other tools (#10858)
### Problem

The first error in #10855 is that when we search a search path which does not include `bash`, lookups fail silently because the discovery script fails to start. The usecase described on #9760 (setting `interpreter_search_path=["<PYENV>"]`) is one example where `bash` will not be present.

### Solution

Recurse to locate an absolute path for `bash` before using `bash`. Additionally, change the "discovery" portion of `BinaryPaths` from `FallibleProcessResult` to `ProcessResult` to fail more quickly in cases where discovery errors, as opposed to succeeding with an empty result.

### Result

The added test passes, and the usecase from #9760 works on my machine.

[ci skip-rust]
[ci skip-build-wheels]