Add `RuleRunner.set_options()` (#10859)
As of https://github.com/pantsbuild/pants/pull/10827, options are now set up through a new `SessionValues` mechanism. This means that our tests should no longer be passing `OptionsBootstrapper` as a `Param` anymore.

Instead, we must mutate the `SchedulerSession` when setting options. We surface a new entry-point `.set_options()`, which will change the options for the life of the `RuleRunner` instance.

```python
rule_runner.set_options(["--pants-ignore=['foo']"])
```

`RuleRunner` already was high on mutation, such as `rule_runner.create_file()` mutating the build root. This is why we encourage using Pytest fixtures to ensure each test is a new instance.

A followup will move `PantsEnvironment` to a `SessionValue`.

[ci skip-rust]
[ci skip-build-wheels]