Remove .*/ from --build-ignore defaults (#10856)
### Problem

We currently have `.*/` as a default glob in the defaults for the `--build-ignore` option; that is, by default ignoring all directories beginning with a dot. This, confusingly, leads to certain situations where it's possible to un-ignore a directory beginning with a dot in the `--pants-ignore` option, and still have that directory fail to appear as a possible target because of the `--build-ignore` default.

### Solution

This commit  removes ".*/" as a default glob in `--build-ignore`, leaving that glob as an overridable default for `--pants-ignore`. It also removes the other defaults for `--build-ignore`, since these pertain to standard directory names for tooling pants doesn't support at this time.

### Result

Explicit overrides of the ".*/" default for `--pants-ignore` now work as intended.