A new mechanism for showing "did you mean" help.
Makes suggestions for unknown goals and unknown flags.

The old system for suggesting similar flag names was
complex, and didn't handle or display scopes in a
user-helpful way. So for example if the user used a
global flag in goal scope, we would show a suggestion
like: "Unknown flag --foobar, did you mean --foobar".

This new implementation handles that case, and moves
the error message computation and display out of the
options system and into the local_pants_runner.py,
which has enough context to do so.

[ci skip-rust]

[ci skip-build-wheels]