Add back `blocked` flag on BoundedCommandRunner workunit (#10873)
### Problem

We're seeing some cases where lines are appearing in the dynamic UI corresponding to the outer workunit in the `BoundedCommandRunner`, that is, the workunit for a process that has been scheduled and is waiting for the opportunity to run. This is problematic because the timer representing how long a workunit has been running for isn't useful in this case. We only want to show workunits representing an actively-running process (or in general an actively-running Pants rule or intrinsic).

The `blocked` flag on `WorkunitMetadata` was added specifically to handle this case, but for some reason it was removed from the metadata for the outer workunit in this code.

### Solution

Add back the `blocked` flag for this outer workunit along with a clarifying comment, and also add the text "(Waiting)" to the description of this workunit, to make it more obvious what's going on if this flag is removed.