Remove the `Scm` interface (#10884)
Right now, our `--changed-since` code uses non-engine code to execute before any rules are used. But we want to be able to port this code to the engine; for example, some users want to run Git for their `setup-py` plugin, and it would be nice if we gave them a safe abstraction pre-built for them.

We weren't using the `Scm` interface; the only actual implementer was `Git`. By removing the interface, we simplify things to then make it easier to port to the engine.

This also renames the folder `scm` to `vcs`. They generally mean the same thing, but SCM has multiple different definitions and can be ambiguous: https://stackoverflow.com/questions/4127425/whats-the-difference-between-vcs-and-scm.

[ci skip-rust]
[ci skip-build-wheels]