Always use the workunit name in stacktraces (#10882)
For stacktraces, it's not very helpful to use the `desc`. We instead want to always use the module name and rule name, which is more helpful for debugging.

This also changes the workunit name for `@goal_rule`s to default to their rule name, rather than the goal's name. This makes things more consistent with other rules. We still special case the `desc` code so that it renders that it's a goal.

Before:

```
Engine traceback:
  in select
  in `list` goal
  in pants.engine.internals.build_files.strip_address_origins
  in Find targets from input specs
  in pants.engine.internals.build_files.addresses_with_origins_from_address_specs
  in pants.engine.internals.build_files.resolve_address
```

After:

```
Engine traceback:
  in select
  in pants.backend.project_info.list_targets.list_targets
  in pants.engine.internals.build_files.strip_address_origins
  in pants.engine.internals.graph.resolve_addresses_with_origins
  in pants.engine.internals.build_files.addresses_with_origins_from_address_specs
  in pants.engine.internals.build_files.resolve_address
```

[ci skip-build-wheels]