Deprecate `awslambda` and `binary` goals in favor of new `package` goal (#10881)
We now have several goals that build an asset:

* `binary`
* `awslambda`
* `setup-py`

Soon, we will be adding support for an `archive()` target type as part of https://github.com/pantsbuild/pants/issues/10733. An `archive` is not really a `binary`, nor any of the other two. We don't want to add a _fourth_ goal `archive`. It gets confusing when there are lots of goals; and it also makes things like `runtime_binary_dependencies` less flexible, e.g. not working with `python_awslambda`.

Semantically, the `package` goal is for transforming your code into a single exportable asset that goes into `--distdir`. This is different than, say, a `compile` goal; running on a library target should no-op.

This only deprecates `awslambda` and `binary` for now; `setup-py` will be a followup, as that doesn't port as easily due to passthrough args.

See https://docs.google.com/document/d/1q77XvcIU9dCzuzA2iFRQQ9AGIfV9jpR269415HdEt0E/edit# for a design doc diving more into this.

[ci skip-rust]
[ci skip-build-wheels]