Add `output_path` field to `python_binary`, `python_awslambda`, and `archive` (#10899)
### Problem

When building an `archive`, we believe users will want to be able to control where certain files/packages show up. This is why we added `relocated_files` in https://github.com/pantsbuild/pants/pull/10880. 

However, `relocated_files` would not work with a package. If you tried using `relocated_files` with a package, then it would not build the actual package, as it's not `FilesSources`.

Further, even if you're not using `archive`, users may want to control the output path when running `./pants package` or using `runtime_package_dependencies` in `python_tests`. For example, they may want to hardcode a certain value so that changing the target name or directory path would not change the final package name.

In v1, we had `basename` for this. But `basename` is not adequate because this solely changes the final file name, but not the full path, like `src.python.pants/pants.pex`.

### Solution

For package target types, allow users to override the default path.

We warn that this can result in ambiguous paths if the user is not careful, whereas our default is always unambiguous. While this could be surprising, the user must go out of their way to opt-in, and we will warn both in `./pants target-types` and the online docs.

[ci skip-build-wheels]
[ci skip-rust]