fs.Digest is declared in Rust (#10905)
### Problem

Our fundamental filesystem types are declared in Python, but declaring them in Rust would cut down on FFI costs, and allow for better error messages by removing some usage of the `externs::project_*` functions.

### Solution

Port `fs.Digest` to Rust. An advantage in the case of `Digest` is that unless the Python side calls `Digest.fingerprint`, we'll never bother to encode the fingerprint as hex (it stays binary in the private `hashing::Digest` instance. And likewise, if the engine is given back a `Digest` object that it had previously given out, it doesn't have to decode hex at the boundary. While it would have been possible to do this with Python code (by having a private `bytes` field for the fingerprint rather than a string), it would have involved redundant logic for validation.

Additionally adds initial Python typestubs for the `native_engine` module, which will likely be the eventual home for documentation on native types.

### Result

For 100 runs:
```
before:
    real    22m40.378s
    user    24m58.706s
    sys     20m28.889s

after:
    real    20m37.329s
    user    23m0.305s
    sys     17m34.746s
```