Port nailgun client to rust (#10865)
This commit ports the nailgun client used in `remote_pants_runner.py` from Python to Rust, using the `nails` crate. This is in preparation for using the Rust nailgun library's heartbeat functionality to replace the current UNIX signal-based mechanism for the client to inform pantsd that a given pants run has been stopped, although this commit does not yet remove that signal-based code.

Removing the Python version of the client allows a lot of Python code around the client side of the nailgun protocol to be removed, and also results in the removal of the logic for retrying connecting to the server on failure. Since a lot of this code is old, and predates the current design of pantsd, I think it makes sense to remove this code now, and add back any retry logic we see as still necessary in future commits.