Send SIGTERM to child processes upon SIGINT (#10930)
### Problem

Currently if a user hits Ctrl-C while at the Python REPL spawned by `./pants repl` on Python targets, pants will quit but the repl process will remain running, listening to stdin on the terminal pants just exited in some kind of indeterminate state, which screws up further input into that terminal.

### Solution

The problem was that when pants receives a SIGINT signal from the user hitting Ctrl-C, it sends SIGINT to all child processes. However, SIGINT does not actually terminate a Python REPL - it prints the text "KeyboardInterrupt" and continues to run. In order to actually shut down the Python process (or any other process that has similar behavior with respect to SIGINT), we need to send the SIGTERM signal instead.

### Result

Hitting Ctrl-C when at the Python REPL now correctly terminates all of pants, pantsd, and the underlying Python REPL process.