Fix some dependencies-like fields not showing up with project introspection (#10923)
### Problem

Closes https://github.com/pantsbuild/pants/issues/10888.

We now have several fields that act like dependencies, but are not quite it. In most code, we don't want the dependencies showing up with `TransitiveTargets` and `DependenciesRequest`; for example, with `relocated_files()`, it's important that the original files targets are not included when we're hydrating sources for the transitive closure.

But with project introspection, we do need these dependencies-like fields to be used. This is important for `--changed-since --changed-dependees` to work properly.

### Solution

Add a new `SpecialCasedDependencies` superclass. This is not a normal field, e.g. not setting `alias`. It's closer in spirit to a "Field template", e.g. `StringField` and `BoolField`.

Both `DependenciesRequest` and `TransitiveTargetsRequest` get new boolean flags to determine if they should include any subclasses of `SpecialCasedDependencies` or not. Then, we teach the project introspection goals to use the new behavior.

[ci skip-rust]