Use cpython types in Rust functions that manipulate python objects (#10942)
### Problem

We have a number of Rust functions in the `externs` module that manipulate Python values that make use of the `Value` type, which is our own reference-counted wrapper around the `PyObject` type provided by the `cpython` library. The `Value` type predates the use of `cpython` for interacting with Python code, and `PyObject` is a more flexible type than the previous type(s) used to interact with Python. Now it is often the case that we take `Value`s returned from one of the `externs` functions and immediately turn it back into a `PyObject`. This makes the API more confusing to reason about, and forces unnecessary and potentially costly type conversions.

### Solution

Implement `AsRef<PyObject>` for `Value` so we can cheaply turn existing parameters of type `Value` into `&PyObject`s. Then, switch out as many function signatures as possible to work with `PyObject`s, using the `AsRef` and other existing conversion traits to turn `PyObject`s into `Value`s where necessary, and lets us save type conversions when we don't need them.