Expose getattr method for Python-related APIs (#10953)
Code that needs to deal with Python objects should be able to use the type-polymorphic `getattr` function exposed in the `externs` mod (which itself wraps the cpython `.getattr()` method on `PyObject`), rather than creating a bespoke `project_<some-type>` method in that module for any particular type they need. This commit makes `getattr` public, which lets us get rid of some `project_` methods and will allow us to get rid of more in future commits. It also has the benefit of pushing the `.unwrap()` calls which were previously hidden in the `project_` methods to the call site, where we can more easily transform them into proper, non-panicking error handling if and when it proves necessary.