Remove tokio Handle type from Executor::new (#10980)
### Problem

Version 0.3 of the `tokio` library no longer has a type called `Handle`, instead unifying the functionality of this type with the `Runtime` type. We currently use the `Handle` type in a bunch of places, particularly tests, where we create our own wrapper around the tokio executor, and we will have to change this in order to upgrade to tokio 0.3.

### Solution

Nearly every call site that invokes the `Executor::new` method passes in `Handle::current`, so this commit modifies that `new` associated function to call `Handle::current` internally, which lets us remove the reference to `Handle` at each call site. This will make it easier to upgrade to tokio 0.3 in the near future.