refactor Core::new including command runner setup (#10993)
### Problem

Core::new is a very convoluted function and it is hard to read what is actually going on in it.

### Solution

Refactor the store and command runner setup into separate functions. (Also refactors the load of some certificates.) This refactor will make the subsequent introduction of remote caching in https://github.com/pantsbuild/pants/pull/10960 easier to comprehend.

### Result

Existing tests pass.