Increase default `[python-setup].resolver_jobs` to `cpu_count / 2` (#11006)
This option controls the concurrency of each Pex subprocess. 

We had set the default low to 2 because—before we added the single resolve optimization—it was common to build multiple PEXes at the same time. This resulted in exhaustion / an OOM issue for Toolchain, which was tricky to figure out.

Now, the user normally builds no more than 1-3 PEXes at a time, unless running `./pants package ::` or does not use a constraints file so we can't use a single resolve. So, our default is missing out on performance.

PEX's normal default is `cpu_count`. Instead, we use `cpu_count / 2` to try to balance performance vs. safety.

[ci skip-rust]
[ci skip-build-wheels]