Restore EngineAware metadata() method with non-cpython-reliant data structures (#11054)
### Problem

It was necessary to revert https://github.com/pantsbuild/pants/pull/11047 , which added a new EngineAware method `metadata()`, since this introduced a dependency on the `cpython` for the `workunit_store` crate, which caused problems in CI.

### Solution

This PR un-reverts the original commit, and modifies the data structures involved so that the `workunit_store` no longer needs to depend on `cpython`. Instead of `WorkunitMetadata` needing to store a (cpython-provided) `PyObject` object, it stores a bespoke `UserMetadataItem` data structure, which simply wraps a UUID. This type is now carried around on `WorkunitMetadata`, and indexes into a per-`Session` hash map storing a map from `UserMetadataItem` to `Value`, which we use when we create Python objects from Workunits. This way, only the engine crate needs to be aware of `cpython` types. 

Additionally, the `PartialEq`, `Eq`, and `Default` implementations for some of the workunit-related types proved no longer necessary, and were removed.