Various Python cleanup refactors  (#11060)
This commit goes through a number of places in the Python parts of the pants codebase where the code is more complex and verbose than it needs to be (presumably because of complexity in previous versions of pants that has since been removed), and simplifies that code. Also adds type annotations in a number of places.