Add process return value as workunit metadata item (#11086)
### Problem

Now that we have the `metadata` item on workunits, we want to be able to have certain intrinsics add their own metadata to worknits. In this case, we want the intrinsic that runs `Process`es to add a metadata item reflecting the exit code.

### Solution

This was a little bit tricky to code up without introducing the `Value` type to sub-crates of the main engine crate. The `UserMetadataItem` type has been split into an enum, where one variant is the wrapped UUID type that lets the workunit type pass around references to Python `Value`s in a session-specific HashMap; the other variant being an "immediate" signed integer. Now, the `process_execution` crate can place a signed integer value in the workunit metadata, and does so with the return value of the process.