Include a host fingerprint in pantsd's identity (#11092)
### Problem

`pantsd` uses only options values to decide which instances are potentially valid for use. But on a well-configured system, option values will be stable across virtual machines and docker images, which means that `pantsd` will fruitlessly attempt to connect to a pid/socket that exists only on a different host. See #10847.

### Solution

In two commits: clean up some unnecessary abstraction around process metadata management, and include a host fingerprint in process metadata.

### Result

Multiple docker or virtual machine instances running in the same workspace should ignore one another's instances of `pantsd`. If we're able to include the uptime in the fingerprint in the future, it would further strengthen this check, but the fact that `docker` generates a random hostname means that this is sufficient to fix #10847.