Remove the examples/ directory.
We now have example repos instead.

A few tests were using some example code, so it was moved
under testprojects/ for now.

Also applies a couple of unrelated style tweaks, and adds
a python_requirement_library for the pycharm remote debugging
client, to make remote debugging easier, as explained in
https://www.pantsbuild.org/docs/python-test-goal#running-tests-interactively.

[ci skip-rust]

[ci skip-build-wheels]