Add `./pants py-constraints --summary` for project overview (#11099)
This is intended to give an overview of the current status of your project, and to help with prioritization for things like Python 3. In contrast, `./pants py-constraints` without `--summary` is intended for debugging on the local terminal.

We export as a CSV and expect users to import into Excel or Google Sheets to analyze the data. For example, https://docs.google.com/spreadsheets/d/1B_2WNmXuXCMbhU6h8nwhSV3YdArJZtE_qODYcI2968o/edit?usp=sharing.

We always run over the entire repo, rather than looking at user specs. This simplifies handling of subtargets and differences between file args vs. target args; both types of addresses are included in the analysis.

[ci skip-rust]
[ci skip-build-wheels]