Improve error message for invalid unions in `Get()`s (#11102)
Before:

```
Engine traceback:
  in select
  in pants.core.goals.lint.lint
Traceback (no traceback):
  <pants native internals>
Exception: Type int is not a member of the LintRequest @union ("A union for StyleRequests that should be lintable.

    Subclass and install a member of this type to provide a linter.
    ")
```

After:

```
Engine traceback:
  in select
  in pants.core.goals.lint.lint
Traceback (no traceback):
  <pants native internals>
Exception: Invalid Get. Because the second argument to `Get(LintResults, LintRequest, 1)` is annotated with `@union`, the third argument should be a member of that union. Did you intend to register `UnionRule(LintRequest, int)`? If not, you may be using the wrong type (int) for the third argument.
```

This also drops the ability for union bases to customize the error message, which hasn't been used for some time. We are using unions very differently than a year ago. For example, we now usually introspect `UnionMembership` to proactively find all valid combinations. At least in our own code, a user should never hit a union error based on something they did, unless us developers had an issue in our code. So, we no longer need to worry about the message making sense to end users; we instead want it to make sense to plugin authors. This is the same logic, for example, for why our wording of the error message for `Get(Snapshot, Digest, 1)` is optimized for plugin authors. 

[ci skip-build-wheels]