Fix logs for `test`, `lint`, `fmt`, and `typecheck` to always print, even if cached (#11111)
Closes https://github.com/pantsbuild/pants/issues/10918.

Rather than inventing some new engine mechanism like a `force_log=True` toggle—which would be costly to maintain—we leverage uncachable rules. We use a pattern where the downstream plugins return the data (e.g. `LintResults`), and then we automatically enrich it for the sake of IO (e.g. `EnrichedLintResults`). This ensures that the data computation is memoized, while the IO is forced to rerun. As we were doing before, we can augment the logging with extra info, e.g. using `--test-output` to determine what to log.

### Result

Regardless of how many times you run this command, we get consistent output (modulo inconsistent ordering of logs):

```
▶ ./pants --changed-since=HEAD~1 lint
01:26:44.10 [INFO] Completed: lint - Docformatter succeeded.
01:26:44.21 [INFO] Completed: lint - isort succeeded.
01:26:44.51 [INFO] Completed: lint - Flake8 succeeded.
01:26:44.75 [INFO] Completed: lint - Black succeeded.
All done! ✨ 🍰 ✨
8 files would be left unchanged.
```

[ci skip-rust]
[ci skip-build-wheels]