Add a specific error for target names containing slashes (#11115)
Having slashes (`/`) in target names used to be implicitly supported, but the new file address syntax in `2.0.x` makes having a slash in a target name ambiguous:
```
this/is/a/file.txt:../../with/a/target
```

We believe that it should be possible to safely rename all targets with slashes in their names, and we think that the file address syntax is worth keeping. So this change deprecates including the relevant characters in target names.

Fixes #11106.

[ci skip-rust]
[ci skip-build-wheels]