remoting: allow timeout parameters to be configured (#11125)
### Problem

The remote store code used in remote execution and remote caching is currently hard-coded with a 10 millisecond timeout that does not increase between retry attempts. While 10 ms may be fine for an on-premises data center, it does not work in less-than-ideal environments (i.e., most computing environments), especially not in AWS.

### Solution

Make the timeouts for remote store RPCs configurable via the new `--remote-store-initial-timeout`, `--remote-store-timeout-multiplier`, and `--remote-store-maximum-timeout` options.

### Result

Existing tests pass.