Re-land the port of Pants' nailgun client to Rust (#11147)
### Problem

#10865 previously landed to port Pants' nailgun client to Rust. It was reverted in #10929 due to an issue with TTY access, where (in particular), the `repl` goal was mostly unresponsive.

### Solution

The unresponsive `repl` was due to a bug in the `nails` library, where `stdin` was being consumed eagerly regardless of whether the server signaled that it would like to receive `stdin`. Pants sends an environment variable to the server that indicates which TTY the client is connected to, and the server will directly connect to that TTY if it can. When the server directly connects to the client's TTY, it does not accept `stdin`, but since `stdin` was read eagerly by the `nails` client (and ending up stuck in a buffer, since the server would not request it), the result was two different processes reading `stdin` from the TTY: the client, and the server.

Bump to `nails` `0.7.0`, which [makes `stdin` initialization lazy](https://github.com/stuhood/nails/commit/6b8c19ae42e3eec29af0cd2af553356344adaf2b).