Deprecate the `sources` field for `python_awslambda` (#11176)
We missed this when making files the atomic unit. It was possible to have >1 file in the `sources` field, which would cause `./pants package` to try building the same lambda >1 time, which would result in a `MergeDigests` error. We knew about this problem with `setup_py()` and fixed it by adding `python_distribution`, but didn't catch this.

Generally, we don't want the `sources` field for this target. It's too common that a `python_library()` uses default sources, but then a `python_awslambda` uses explicit sources for a file, and there end up with >1 owner for that file. This turns off dep inference for that file, which is extremely confusing. We instead follow the `python_distribution` model of no `sources`.

Deprecations:

```
18:24:16.43 [WARN] /Users/eric/DocsLocal/code/projects/pants/src/python/pants/engine/target.py:226: DeprecationWarning: DEPRECATED: the 'sources' field will be removed in version 2.2.0.dev0.
  Using the `sources` field in the target src/python/pants/backend/awslambda/python:lambda. Remove the `sources` field and create a new `python_library()` target (if you do not yet have one), then add the `python_library()` to the `dependencies` field of this `python_awslambda`. See https://www.pantsbuild.org/docs/awslambda-python for an example.
  super().__init__(raw_value, address=address)
18:24:16.43 [WARN] /Users/eric/DocsLocal/code/projects/pants/src/python/pants/engine/target.py:134: DeprecationWarning: DEPRECATED: the 'compatibility' field will be removed in version 2.2.0.dev0.
  Using the `compatibility` field in the target src/python/pants/backend/awslambda/python:lambda. Because the `sources` field will be removed, it no longer makes sense to have a `compatibility` field for `python_awslambda` targets. Instead, set the `interpreter_constraints` field on the `python_library` target containing this Lambda's handler code.
  super().__init__(raw_value, address=address)
```

[ci skip-rust]
[ci skip-build-wheels]